Remove `MailApi` and move its content to `MailData`
